backend/requirements.txt
# File: backend/requirements.txt

# Web framework for creating the API
fastapi
uvicorn[standard]

# AI and Data processing libraries
pinecone
langchain
langchain-text-splitters
google-generativeai
langchain-google-genai
cohere

#For loading settongs from .env file
pydantic-settings
